pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dpette"
version = "0.1.0"
description = "Reverse-engineered USB/serial driver for DLAB dPette electronic pipettes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-mock>=3.12,<4",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "types-pyserial>=3.5",
    "pre-commit>=3.7,<4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dpette"]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"tools/*" = ["E501", "B007", "SIM115"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tools.*"
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
